nftables.service: load firewall earlier in the boot process
The systemd unit file /lib/systemd/system/nftables.service should contain
a Before=network-pre.target statement, so that firewall rules are
established before network interfaces are up. Otherwise, there may be a
small time window during which the host is not protected by the firewall.
This is also recommended in the systemd documentation; see
https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

Furthermore, the unit should be WantedBy=network.target instead of
WantedBy=multi-user.target, so that it is started when the network is
started, even if the system is not in multi-user mode.

Closes: #866902
Signed-off-by: Martin Dickopp <martin@zero-based.org>
Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>