upstream: fix d/patches changes applied to the upstream tree
I accidentally commited the changes of debian/patches/* to the upstream
tree in the master branch. This is a mistake, revert the changes now.

The upstream code should be keep as if it was deployed right from the upstream
tarball, ready to apply our patches on top.

Since the source package is composed of the upstream tarball + the debian
tarball (which contains just debian/), this doesn't affect the package in the
archive and only the git repo itself.

Git-Dch: Ignore
Signed-off-by: Arturo Borrero Gonzalez <arturo@debian.org>