d/rules: fix non-reproducible dates in man-pages
asciidoc embeds a revision date into the man-pages it creates.  If not explicit-
ly set, the date used is the date when the page is generated.  This is not re-
producible and can result in different dates in packages for different archi-
tectures, which interferes with multi-arch hinting.  In order to help make its
output reproducible, asciidoc checks whether the `SOURCE_DATE_EPOCH` environment
variable is set and uses the value of that to derive the date if it is.  Include
/usr/share/dpkg/pkg-info.mk to make the variable available.

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>